Clean up old files and codes
