+++ frontend/src/components/RecordsList.tsx
@@ -1,81 +1,69 @@
 import AutoSizer from "react-virtualized-auto-sizer";
 import { Button, Container, Modal } from "react-bootstrap";
-import { Dispatch, SetStateAction, useCallback, useMemo, useRef } from "react";
+import { Dispatch, SetStateAction, useCallback, useRef } from "react";
 import { useNavigate, useParams } from "react-router-dom";
 import { DownloadModal } from "./DownloadModal";
-import { buildTsvString } from "../utils/stringBuilders";
+import {
+  buildSentenceCaseString,
+  buildTsvString,
+} from "../utils/stringBuilders";
 import { AgGridReact } from "ag-grid-react";
 import { useState } from "react";
 import styles from "./records.module.scss";
 import "ag-grid-community/styles/ag-grid.css";
 import "ag-grid-community/styles/ag-theme-alpine.css";
 import "ag-grid-enterprise";
-import {
-  ColDef,
-  IServerSideDatasource,
-  IServerSideGetRowsParams,
-  IServerSideGetRowsRequest,
-} from "ag-grid-community";
+import { ColDef, IServerSideGetRowsParams } from "ag-grid-community";
 import { DataName, useHookLazyGeneric } from "../shared/types";
 import SamplesList, { SampleContext } from "./SamplesList";
-import { PatientIdsTriplet, SortDirection } from "../generated/graphql";
-import { defaultColDef } from "../shared/helpers";
 import {
-  ErrorMessage,
-  LoadingSpinner,
-  Toolbar,
-} from "../shared/tableElements-records";
+  DashboardRecordFilter,
+  DashboardRecordSort,
+  PatientIdsTriplet,
+  QueryDashboardCohortsArgs,
+  QueryDashboardPatientsArgs,
+  QueryDashboardRequestsArgs,
+} from "../generated/graphql";
+import { defaultColDef } from "../shared/helpers";
+import { ErrorMessage, Toolbar } from "../shared/tableElements";
 import { AgGridReact as AgGridReactType } from "ag-grid-react/lib/agGridReact";
 import { BreadCrumb } from "../shared/components/BreadCrumb";
 import { Title } from "../shared/components/Title";
+import { parseUserSearchVal } from "../utils/parseSearchQueries";
 
-const CACHE_BLOCK_SIZE = 100;
+const CACHE_BLOCK_SIZE = 100; // number of rows to fetch at a time
+const MAX_ROWS_EXPORT = 10000;
 
-interface IRecordsListProps {
-  colDefs: ColDef[];
+interface INewRecordsListProps {
+  columnDefs: ColDef[];
   dataName: DataName;
   enableInfiniteScroll?: boolean;
   lazyRecordsQuery: typeof useHookLazyGeneric;
-  lazyRecordsQueryAddlVariables?: Record<string, any>;
-  prepareDataForAgGrid?: (
-    data: any,
-    filterModel: IServerSideGetRowsRequest["filterModel"],
-    sortModel?: { [key: string]: SortDirection }[] | undefined
-  ) => any;
-  queryFilterWhereVariables: (
-    parsedSearchVals: string[]
-  ) => Record<string, any>[];
-  defaultSort?: { [key: string]: SortDirection }[];
+  defaultSort: DashboardRecordSort;
   userSearchVal: string;
   setUserSearchVal: Dispatch<SetStateAction<string>>;
-  parsedSearchVals: string[];
-  setParsedSearchVals: Dispatch<SetStateAction<string[]>>;
-  handleSearch: () => void;
+  setCustomSearchStates?: Dispatch<SetStateAction<PatientIdsTriplet[]>>;
+  searchInterceptor?: (userSearchVal: string) => Promise<string[]>;
   showDownloadModal: boolean;
   setShowDownloadModal: Dispatch<SetStateAction<boolean>>;
-  handleDownload: () => void;
+  handleDownload: (recordCount: number) => void;
   samplesColDefs: ColDef[];
   sampleContext?: SampleContext;
   userEmail?: string | null;
   setUserEmail?: Dispatch<SetStateAction<string | null>>;
-  setCustomSearchVals?: Dispatch<SetStateAction<PatientIdsTriplet[]>>;
   customToolbarUI?: JSX.Element;
 }
 
 export default function RecordsList({
-  colDefs,
+  columnDefs,
   dataName,
   enableInfiniteScroll = true,
   lazyRecordsQuery,
-  lazyRecordsQueryAddlVariables,
-  prepareDataForAgGrid,
-  queryFilterWhereVariables,
   defaultSort,
   userSearchVal,
   setUserSearchVal,
-  parsedSearchVals,
-  setParsedSearchVals,
-  handleSearch,
+  setCustomSearchStates,
+  searchInterceptor,
   showDownloadModal,
   setShowDownloadModal,
   handleDownload,
@@ -84,93 +72,95 @@ export default function RecordsList({
   userEmail,
   setUserEmail,
   customToolbarUI,
-  setCustomSearchVals,
-}: IRecordsListProps) {
+}: INewRecordsListProps) {
   const [showClosingWarning, setShowClosingWarning] = useState(false);
   const [unsavedChanges, setUnsavedChanges] = useState(false);
-  const [filterModel, setFilterModel] = useState<Record<string, any>>({});
-  const [rowCount, setRowCount] = useState<number>(0);
-  const [uniqueSampleCount, setUniqueSampleCount] = useState<number>(0);
 
   const gridRef = useRef<AgGridReactType>(null);
   const navigate = useNavigate();
   const params = useParams();
 
-  // note that we aren't using initial fetch
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  const [initialFetch, { loading, error, data, fetchMore, refetch }] =
-    lazyRecordsQuery({
-      variables: {
-        options: {
-          limit: CACHE_BLOCK_SIZE,
-          offset: 0,
-          sort: defaultSort,
-        },
-        ...lazyRecordsQueryAddlVariables,
-      },
-    });
+  const [, { error, data, fetchMore, refetch }] = lazyRecordsQuery({
+    variables: {
+      searchVals: [],
+      sort: defaultSort,
+      limit: CACHE_BLOCK_SIZE,
+      offset: 0,
+    },
+  });
 
-  const totalCountNodeName = `${dataName}Connection`;
+  const dataNameCapitalized = buildSentenceCaseString(dataName);
+  const recordsQueryName = `dashboard${dataNameCapitalized}`;
+  const recordCountQueryName = `dashboard${dataNameCapitalized.slice(
+    0,
+    -1
+  )}Count`;
+  const recordCount = data?.[recordCountQueryName]?.totalCount;
+  const uniqueSampleCount = data?.[recordCountQueryName]?.uniqueSampleCount;
 
-  const getSortModel = useCallback(() => {
-    const sortModel = gridRef.current?.columnApi
-      ?.getColumnState()
-      .filter(({ sort }) => sort)
-      .map(({ colId, sort }) => ({
-        [colId]: sort?.toUpperCase() as SortDirection,
-      }));
-    if (sortModel && sortModel.length > 0) {
-      return sortModel;
-    }
-    return defaultSort;
-  }, [defaultSort]);
+  const getServerSideDatasource = useCallback(
+    ({ searchVals }) => {
+      return {
+        getRows: async (params: IServerSideGetRowsParams) => {
+          let filters: DashboardRecordFilter[] | undefined;
+          const filterModel = params.request.filterModel;
+          if (filterModel && Object.keys(filterModel).length > 0) {
+            filters = Object.entries(filterModel).map(([key, value]) => ({
+              field: key,
+              // Flexibly handle AG Grid's `any` type for filter settings by JSON.parse() this string value,
+              // then check the field name before consuming it at the GraphQL server (see https://stackoverflow.com/a/45601881)
+              filter: JSON.stringify(value),
+            }));
+          } else {
+            filters = undefined; // all filter values are selected
+          }
 
-  const datasource: IServerSideDatasource = useMemo(() => {
-    return {
-      // called by the grid when more rows are required
-      getRows: (params: IServerSideGetRowsParams) => {
-        const fetchInput = {
-          where: {
-            OR: queryFilterWhereVariables(parsedSearchVals),
-          },
-          options: {
-            offset: params.request.startRow,
-            limit: params.request.endRow,
-            sort: getSortModel(),
-          },
-        };
+          const fetchInput = {
+            searchVals,
+            sort: params.request.sortModel[0] || defaultSort,
+            filters,
+            offset: params.request.startRow ?? 0,
+            limit: CACHE_BLOCK_SIZE,
+          } as
+            | QueryDashboardRequestsArgs
+            | QueryDashboardPatientsArgs
+            | QueryDashboardCohortsArgs;
 
-        // if this is NOT first call, use refetch
-        // (which is analogous in this case to the original fetch)
-        const thisFetch =
-          params.request.startRow! === 0
-            ? refetch(fetchInput)
-            : fetchMore({
-                variables: fetchInput,
+          const thisFetch =
+            params.request.startRow === 0
+              ? refetch(fetchInput)
+              : fetchMore({
+                  variables: fetchInput,
+                });
+
+          return thisFetch
+            .then((result) => {
+              params.success({
+                rowData: result.data[recordsQueryName],
+                rowCount: result.data[recordCountQueryName].totalCount,
               });
+            })
+            .catch((error) => {
+              console.error(error);
+              params.fail();
+            });
+        },
+      };
+    },
+    [defaultSort, refetch, fetchMore, recordsQueryName, recordCountQueryName]
+  );
 
-        return thisFetch.then((d) => {
-          let agGridData = d.data;
-          if (prepareDataForAgGrid) {
-            agGridData = prepareDataForAgGrid(d.data, filterModel);
-          }
-
-          if ("uniqueSampleCount" in agGridData) {
-            setUniqueSampleCount(agGridData.uniqueSampleCount);
-          }
-          setRowCount(agGridData[totalCountNodeName].totalCount);
-
-          params.success({
-            rowData: agGridData[dataName],
-            rowCount: agGridData[totalCountNodeName].totalCount,
-          });
-        });
-      },
-    };
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [parsedSearchVals, filterModel]);
-
-  if (loading) return <LoadingSpinner />;
+  async function refreshData(userSearchVal: string) {
+    const extraSearchVals = searchInterceptor
+      ? await searchInterceptor(userSearchVal)
+      : [];
+    const searchVals = [
+      ...parseUserSearchVal(userSearchVal),
+      ...extraSearchVals,
+    ];
+    const newDatasource = getServerSideDatasource({ searchVals });
+    gridRef.current?.api.setServerSideDatasource(newDatasource); // triggers a refresh
+  }
 
   if (error) return <ErrorMessage error={error} />;
 
@@ -190,29 +180,16 @@ export default function RecordsList({
       {showDownloadModal && (
         <DownloadModal
           loader={async () => {
-            const sortModel = getSortModel();
-
+            // Using fetchMore instead of refetch to avoid overriding the cached variables
             const { data } = await fetchMore({
               variables: {
-                where: {
-                  OR: queryFilterWhereVariables(parsedSearchVals),
-                },
-                options: {
-                  offset: 0,
-                  limit: undefined,
-                  sort: sortModel,
-                },
+                offset: 0,
+                limit: MAX_ROWS_EXPORT,
               },
             });
-
-            let agGridData = data;
-            if (prepareDataForAgGrid) {
-              agGridData = prepareDataForAgGrid(data, filterModel, sortModel);
-            }
-
             return buildTsvString(
-              agGridData[dataName],
-              colDefs,
+              data[recordsQueryName],
+              columnDefs,
               gridRef.current?.columnApi?.getAllGridColumns()
             );
           }}
@@ -285,21 +262,16 @@ export default function RecordsList({
         dataName={dataName}
         userSearchVal={userSearchVal}
         setUserSearchVal={setUserSearchVal}
-        handleSearch={handleSearch}
-        clearUserSearchVal={() => {
-          setCustomSearchVals && setCustomSearchVals([]);
-          setParsedSearchVals([]);
-        }}
-        matchingResultsCount={`${rowCount?.toLocaleString()} matching ${
-          rowCount !== 1 ? dataName : dataName.slice(0, -1)
-        }${
-          enableInfiniteScroll
-            ? ""
-            : ` (${uniqueSampleCount.toLocaleString()} unique ${
-                uniqueSampleCount !== 1 ? "samples" : "sample"
-              })`
+        setCustomSearchStates={setCustomSearchStates}
+        onSearch={async (userSearchVal) => refreshData(userSearchVal)}
+        matchingResultsCount={`${
+          recordCount !== undefined ? recordCount.toLocaleString() : "Loading"
+        } matching ${dataName} ${
+          uniqueSampleCount
+            ? `(${uniqueSampleCount.toLocaleString()} unique samples)`
+            : ""
         }`}
-        handleDownload={handleDownload}
+        onDownload={() => handleDownload(recordCount)}
         customUIRight={customToolbarUI}
       />
 
@@ -311,11 +283,10 @@ export default function RecordsList({
           >
             <AgGridReact
               ref={gridRef}
-              rowModelType={"serverSide"}
-              columnDefs={colDefs}
-              serverSideDatasource={datasource}
+              rowModelType="serverSide"
               serverSideInfiniteScroll={enableInfiniteScroll}
               cacheBlockSize={CACHE_BLOCK_SIZE}
+              columnDefs={columnDefs}
               debug={false}
               context={{
                 navigateFunction: navigate,
@@ -328,9 +299,7 @@ export default function RecordsList({
                 params.columnApi.autoSizeAllColumns();
               }}
               enableRangeSelection={true}
-              onFilterChanged={(params) => {
-                setFilterModel(params.api.getFilterModel());
-              }}
+              onGridColumnsChanged={() => refreshData(userSearchVal)}
             />
           </div>
         )}
