+++ frontend/src/shared/types.ts
@@ -1,5 +1,5 @@
 import * as Apollo from "@apollo/client";
-import { RequestsListDocument } from "../generated/graphql";
+import { DashboardRequestsDocument } from "../generated/graphql";
 
 const defaultOptions = {} as const;
 
@@ -7,7 +7,7 @@ export function useHookLazyGeneric(
   baseOptions?: Apollo.LazyQueryHookOptions<any, any>
 ) {
   const options = { ...defaultOptions, ...baseOptions };
-  return Apollo.useLazyQuery<any, any>(RequestsListDocument, options);
+  return Apollo.useLazyQuery<any, any>(DashboardRequestsDocument, options);
 }
 
 export type DataName = "requests" | "patients" | "cohorts" | "samples";
