+++ graphql/operations.graphql
@@ -1,51 +1,3 @@
-query RequestsList($options: RequestOptions, $where: RequestWhere) {
-  requestsConnection(where: $where) {
-    totalCount
-  }
-  requests(where: $where, options: $options) {
-    ...RequestParts
-    importDate
-    totalSampleCount
-    hasMetadataRequestMetadata {
-      importDate
-    }
-    hasSampleSamplesConnection {
-      totalCount
-    }
-  }
-}
-
-query PatientsList($options: PatientOptions, $where: PatientWhere) {
-  patientsConnection(where: $where) {
-    totalCount
-  }
-  patients(where: $where, options: $options) {
-    smilePatientId
-    cmoPatientId
-    dmpPatientId
-    totalSampleCount
-    cmoSampleIds
-    consentPartA
-    consentPartC
-    hasSampleSamples {
-      smileSampleId
-      hasMetadataSampleMetadata {
-        primaryId
-        cmoSampleName
-        additionalProperties
-        cmoPatientId
-      }
-    }
-    hasSampleSamplesConnection {
-      totalCount
-    }
-    patientAliasesIsAlias {
-      namespace
-      value
-    }
-  }
-}
-
 query DashboardRequests(
   $searchVals: [String!]
   $filters: [DashboardRecordFilter!]
@@ -269,49 +221,5 @@ query GetPatientIdsTriplets($patientIds: [String!]!) {
     CMO_ID
     DMP_ID
     PT_MRN
-  }
-}
-
-query CohortsList(
-  $where: CohortWhere
-  $options: CohortOptions
-  $hasCohortCompleteCohortCompletesOptions2: CohortCompleteOptions
-) {
-  cohortsConnection(where: $where) {
-    totalCount
-  }
-  cohorts(where: $where, options: $options) {
-    cohortId
-    smileSampleIds
-    totalSampleCount
-    billed
-    initialCohortDeliveryDate
-    endUsers
-    pmUsers
-    projectTitle
-    projectSubtitle
-    status
-    type
-    hasCohortCompleteCohortCompletes(
-      options: $hasCohortCompleteCohortCompletesOptions2
-    ) {
-      date
-      endUsers
-      pmUsers
-      projectTitle
-      projectSubtitle
-      status
-      type
-    }
-    hasCohortSampleSamplesConnection {
-      totalCount
-    }
-    hasCohortSampleSamples {
-      smileSampleId
-      hasTempoTempos {
-        smileTempoId
-        billed
-      }
-    }
   }
 }
